+++ HISTORY.rst
@@ -5,9 +5,12 @@ History
 
 As of 1.0.0, `semantic versioning <http://semver.org/>`__ is used.
 
-3.1.1-dev
-+++++
+4.0.0-dev
++++++++++
 released XXXX-XX-XX
+
+- breaking: remove webclient.change_song_metadata; use mobileclient.change_song_metadata instead
+- breaking: webclient.upload_album_art now returns a url to the uploaded image
 
 3.1.0
 +++++